Add functionality to Student Registration form

+++ src/main/java/com/dileep/erp/repository/StudentRepository.java
@@ -3,5 +3,5 @@ package com.dileep.erp.repository;
 import com.dileep.erp.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student,Integer> {
+public interface StudentRepository extends JpaRepository<Student,Long> {
 }
